requirements.txt
yfinance>=0.2.36
pandas>=2.2.0
numpy>=1.26.0
streamlit>=1.31.0
plotly>=5.18.0
fredapi>=0.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
